+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## latest
+## v0.4.0-alpha
+* Renamed:
+  - `Execute` to `ExecuteV1`.
+  - `ExecuteAsyncWait` to `Execute`.
+
+## v0.3.0-alpha
+* Added `CancelTransaction` feature.
+
+## v0.2.0-alpha
 
 * Added v2 predefined results formats:
   - `GetTransactions` returns `TransactionsAsyncMultipleResponses`.